// src/Types/FormType.tsx
export interface Inputs {
  name: string;
  email: string;
  website: string;
  comment: string;
};